+++ internal/schemautil/mutations.go
@@ -1,5 +1,7 @@
 package schemautil
 
+import "golang.org/x/exp/slices"
+
 // sensitiveFields is a list of fields that are not returned by the API on a refresh, but are supposed to remain in the
 // state to make Terraform work properly.
 var sensitiveFields = []string{
@@ -19,109 +21,18 @@ func copySensitiveFields(old, new map[string]interface{}) {
 // normalizeIPFilter compares a list of IP filters set in the old user config and a sorted version coming from the new
 // user config and returns the re-sorted IP filters, such that all matching entries will be in the same order as
 // defined in the old user config.
-func normalizeIPFilter(old, new map[string]interface{}) {
-	oldIPFilters, _ := old["ip_filter"].([]interface{})
-	newIPFilters, _ := new["ip_filter"].([]interface{})
-	fieldToWrite := "ip_filter"
-
-	if oldIPFilters == nil || newIPFilters == nil {
-		var ok bool
-
-		oldIPFilters, _ = old["ip_filter_string"].([]interface{})
-
-		newIPFilters, ok = new["ip_filter_string"].([]interface{})
-
-		fieldToWrite = "ip_filter_string"
-
-		if !ok {
-			oldIPFilters, ok = old["ip_filter_object"].([]interface{})
-			if !ok {
-				return
-			}
-
-			newIPFilters, ok = new["ip_filter_object"].([]interface{})
-			if !ok {
-				return
-			}
-
-			fieldToWrite = "ip_filter_object"
-		}
+func normalizeIPFilter(_, new map[string]interface{}) {
+	if ip, ok := new["ip_filter_object"].([]interface{}); ok {
+		sortObjects(ip, "network")
 	}
 
-	var normalizedIPFilters []interface{}
-	var nonexistentIPFilters []interface{}
-
-	// First, we take all the elements from old and if they match with the elements in new,
-	// we preserve them in the same order as they were defined in old.
-	for _, o := range oldIPFilters {
-		found := false
-
-		for _, n := range newIPFilters {
-			// Define two comparison variables to avoid code duplication in the loop.
-			var comparableO interface{}
-
-			var comparableN interface{}
-
-			// If we're dealing with a string format, we need to compare the values directly.
-			if fieldToWrite == "ip_filter" || fieldToWrite == "ip_filter_string" {
-				comparableO = o
-
-				comparableN = n
-			} else {
-				// If we're dealing with an object format, we need to compare the values of the "network" field.
-				comparableO = o.(map[string]interface{})["network"]
-
-				comparableN = n.(map[string]interface{})["network"]
-			}
-
-			if comparableO == comparableN {
-				normalizedIPFilters = append(normalizedIPFilters, o)
-
-				found = true
-
-				break
-			}
-		}
-
-		// We append old elements that are missing in new to nonexistentIPFilters.
-		if !found {
-			nonexistentIPFilters = append(nonexistentIPFilters, o)
-		}
+	if ip, ok := new["ip_filter_string"].([]interface{}); ok {
+		sortStrings(ip)
 	}
 
-	// Second, we take the new and check whether there are any differences with the old, and
-	// append those to nonexistentIPFilters.
-	for _, n := range newIPFilters {
-		found := false
-
-		for _, o := range oldIPFilters {
-			var comparableO interface{}
-
-			var comparableN interface{}
-
-			if fieldToWrite == "ip_filter" || fieldToWrite == "ip_filter_string" {
-				comparableO = o
-
-				comparableN = n
-			} else {
-				comparableO = o.(map[string]interface{})["network"]
-
-				comparableN = n.(map[string]interface{})["network"]
-			}
-
-			if comparableO == comparableN {
-				found = true
-
-				break
-			}
-		}
-
-		if !found {
-			nonexistentIPFilters = append(nonexistentIPFilters, n)
-		}
+	if ip, ok := new["ip_filter"].([]interface{}); ok {
+		sortStrings(ip)
 	}
-
-	new[fieldToWrite] = append(normalizedIPFilters, nonexistentIPFilters...)
 }
 
 // stringSuffixForIPFilters adds a _string suffix to the IP filters.
@@ -148,4 +59,22 @@ func stringSuffixForNamespaces(new map[string]interface{}) {
 	new["namespace_string"] = namespaces
 
 	new["namespaces"] = nil
+}
+
+// sortStrings sorts untyped list of strings
+func sortStrings(list []interface{}) {
+	slices.SortFunc(list, func(a, b interface{}) bool {
+		aa, _ := a.(string)
+		bb, _ := b.(string)
+		return aa < bb
+	})
+}
+
+// sortObjects sorts list of objects by give key
+func sortObjects(list []interface{}, key string) {
+	slices.SortFunc(list, func(a, b interface{}) bool {
+		aa, _ := a.(map[string]interface{})[key].(string)
+		bb, _ := b.(map[string]interface{})[key].(string)
+		return aa < bb
+	})
 }